update tests (code cleanup)
